Move http proxy fixture
